back to alan branch....oops

+++ src/mustc/be/User.java
@@ -64,7 +64,7 @@ public class User {
         this.password = password;
     }
 
-    public int getSalary() {
+    public float getSalary() {
         return salary;
     }
